+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Bruce lee".
 - Tests for "Ancestors & descendants".
 - Tests for "Paving with bricks".
+- Tests for "Playfair cipher".
 
 ## [1.14.0] - 2022-10-31
 ### Added